Refactor to use a ScriptableRenderFeature for compatibility with URP.
- Removed support for mandlebulb, but should be easy enough to add back in if desired.
- One remaining issue: it RaymarchRenderPassFeature.cs, releasing the
'shapes' ComputeBuffer in the FrameCleanup method (where I would think
is the correct place to release it) breaks the ComputeShader: the
ComputeShader appears to get an all-zeros StructuredBuffer. I've moved
the .dispose() call to just before the re-allocation in Configure(),
but this occasionally produces the warning about an undisposed
ComputeBuffer. Could be a Unity bug? Or just something I'm not
understanding about the order of events when using CommandBuffers.